Split common parts of tests into `test/utils.js`
